+++ modules/documentation/docs/changelog.md
@@ -2,6 +2,7 @@
 
 ## Next Release
 
+- [engine] TRANSFER_ROUTING_COMPLETE event emitted through to original sender.
 - \[router\] Improvements to `autoRebalance` for better state machine. Separate sending tx and waiting for confirmation. Store tx hashes after tx submit.
 
 ## 0.2.4-beta.1